Add support for an aggregation rules LRU cache
A new option, AGGREGATION_RULES_CACHE_SIZE can be
specified (default infinity which turns of LRU)
which limits the aggregation rule resolving cache,
which stores the results rule regex as applied to
an incoming metric.

Set to an integral value to limit the size of the cache.
Uses ordereddict, which is provided with Python 2.7 and
installable for older versions.  If the import fails,
the non-LRU cache is used.